Fixed bug with clustering
closest_point_index = np.argmin(distances)  # index of the closest point

This line of code assumes that the distances array is not empty. However, if all points have already been assigned to clusters and the points list is empty, then the distances array will also be empty, leading to this error.

To fix this issue, we need to add a check to ensure that there are still points left to be assigned to clusters before we try to find the closest point.