Fixed bug with calculating final index
final_index += 1 / cluster_radius  # add the reciprocal of the cluster radius to the final index

This line of code assumes that the cluster radius is not zero. However, if a cluster contains only one point, the cluster radius will be zero, leading to this warning.

To fix this issue, we need to add a check to ensure that the cluster radius is not zero before we try to calculate its reciprocal.