Updated calculate_final_index to fix bug with NDRindex greater than 1
In our Python implementation, this is done with the following lines:

cluster_radius = np.mean(distances)  # average distance, defined as the cluster radius
final_index += 1 / cluster_radius  # add the reciprocal of the cluster radius to the final index

The issue here is that we're adding the reciprocal of the cluster radius to the final index, rather than 1 - (cluster_radius / average_scale). This is why we're getting a value greater than 1.